Allow deletion of tasks, use projects as reference
Tasks can now be deleted.  Projects are no longer
searched for by index, instead a reference is
stored in the HTML elements for later use,
specifically the project list button for loading
a project view and the add ask button for adding
a task to a project.

A reference to the parent project is also now
stored in each task.